Implemented Advanced Search Feature
Added functionality for the advanced search tab with fields hospital, department, and physician. Like the basic search feature, advanced search also filters out the search of a patient as fields are filled in for better user experience.

Matched the scss to the basic search feature. SCSS still needs further improvement overall but implemented for functionality for now.